third_party/blink/web_tests/fast/forms/ValidityState-patternMismatch-unsupported-expected.txt

Check if the pattern constraint is not applied to some input types.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS input.validity.patternMismatch is false
PASS input.validity.patternMismatch is true
PASS successfullyParsed is true

TEST COMPLETE
